+++ user-service/src/main/java/com/flight_booking/user_service/application/service/AuthService.java
@@ -1,13 +1,11 @@
 package com.flight_booking.user_service.application.service;
 
-import com.flight_booking.user_service.domain.model.Role;
 import com.flight_booking.user_service.domain.model.User;
 import com.flight_booking.user_service.domain.repository.UserRepository;
 import com.flight_booking.user_service.infrastructure.security.authentication.CustomUserDetails;
 import com.flight_booking.user_service.infrastructure.security.jwt.JwtUtil;
 import com.flight_booking.user_service.presentation.global.exception.ErrorCode;
 import com.flight_booking.user_service.presentation.global.exception.UserException;
-import com.flight_booking.user_service.presentation.request.ChangePwRequest;
 import com.flight_booking.user_service.presentation.request.FindIdRequest;
 import com.flight_booking.user_service.presentation.request.SignUpRequest;
 import java.util.List;
@@ -17,7 +15,6 @@ import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -61,10 +58,7 @@ public class AuthService {
       CustomUserDetails userDetails = (CustomUserDetails) authentication.getPrincipal();
 
       String validatedEmail = userDetails.getUsername();
-      String role = userDetails.getAuthorities().stream()
-          .map(GrantedAuthority::getAuthority)
-          .findFirst()
-          .orElse(Role.USER.getAuthority());
+      String role = userDetails.getAuthorities().toString();
 
       // 사용자 상태 확인 (블락/ 탈퇴)
       checkUserStatus(validatedEmail);
@@ -106,35 +100,4 @@ public class AuthService {
 
     return user.getEmail();
   }
-
-  // 비밀번호 변경
-  @Transactional
-  public void changePw(String email, ChangePwRequest request) {
-    User user = getUser(email);
-
-    // 기존 비밀번호 검증
-    if (!passwordEncoder.matches(request.currentPw(), user.getPassword())) {
-      throw new UserException(ErrorCode.INVALID_CURRENT_PASSWORD);
-    }
-
-    // 기존 비밀번호와 새로운 비밀번호가 같은지 확인
-    if (passwordEncoder.matches(request.newPw(), user.getPassword())) {
-      throw new UserException(ErrorCode.PASSWORDS_SAME);
-    }
-
-    // 새로운 비밀번호와 확인 비밀번호가 일치하는지 확인
-    if (!request.newPw().equals(request.confirmPw())) {
-      throw new UserException(ErrorCode.PASSWORDS_DO_NOT_MATCH);
-    }
-
-    user.setPassword(passwordEncoder.encode(request.newPw()));
-  }
-
-
-  // 사용자 찾기
-  private User getUser(String email) {
-    return userRepository.findByEmail(email)
-        .orElseThrow(() -> new UserException(ErrorCode.USER_NOT_FOUND));
-  }
-
 }
